Plain CreatesNewUser service.
