--- app/services/creates_new_user/generates_password.rb
@@ -1,14 +1,0 @@
-module Services
-  module Actions
-    class GeneratesPassword
-      extend LightService::Action
-
-      expects :user
-
-      executed do |context|
-        user = context[:user]
-        user.password = Sorcery::Model::TemporaryToken.generate_random_token unless user.internal?
-      end
-    end
-  end
-end
