--- app/services/creates_new_user/sends_activation_needed_email.rb
@@ -1,15 +1,0 @@
-module Services
-  module Actions
-    class SendsActivationNeededEmail
-      extend LightService::Action
-
-      expects :user
-
-      executed do |context|
-        user = context[:user]
-
-        UserMailer.activation_needed_email(user).deliver_now unless user.activated?
-      end
-    end
-  end
-end
